optimizers and decay: code simplification
