+++ neuralmonkey/functions.py
@@ -1,6 +1,10 @@
-from typing import Optional
+"""Collection of various functions and function wrappers."""
 
+from typing import Optional, Any
+
+import math
 import tensorflow as tf
+from typeguard import check_argument_types
 
 
 def inverse_sigmoid_decay(param, rate, min_value: float = 0.,
@@ -50,3 +54,29 @@ def piecewise_function(param, values, changepoints, name=None,
         predicates = [tf.less(param, x) for x in changepoints]
         return tf.case(list(zip(predicates, lambdas[:-1])), lambdas[-1],
                        name=s_name)
+
+
+def noam_decay(learning_rate: float,
+               model_dimension: int,
+               warmup_steps: int) -> tf.Tensor:
+    """Return decay function as defined in Vaswani et al., 2017, Equation 3.
+
+    Arguments:
+        model_dimension: Size of the hidden states of decoder and encoder
+        warmup_steps: Number of warm-up steps
+    """
+    step = tf.to_float(tf.train.get_or_create_global_step())
+
+    inv_sq_dim = 1 / math.sqrt(model_dimension)
+    inv_sq3_warmup_steps = math.pow(warmup_steps, -1.5)
+
+    inv_sq_step = 1 / tf.sqrt(step)
+    warmup = step * inv_sq3_warmup_steps
+
+    return learning_rate * inv_sq_dim * tf.minimum(inv_sq_step, warmup)
+
+
+def variable(initial_value: Any = 0,
+             trainable: bool = False, **kwargs) -> tf.Variable:
+    check_argument_types()
+    return tf.Variable(initial_value, trainable, **kwargs)
