+++ tests/bandit.ini
@@ -84,10 +84,10 @@ class=trainers.generic_trainer.GenericTrainer
 objectives=[<bandit>]
 l2_weight=1.0e-8
 clip_norm=1.0
-optimizer_getter=<lazyadam_g>
+optimizer=<lazyadam_g>
 
 [lazyadam_g]
-class=trainers.optimizers.lazy_adam_optimizer
+class=tf.train.AdamOptimizer
 beta1=0.9
 beta2=0.98
 epsilon=1.0e-9
