+++ tests/str.ini
@@ -107,10 +107,10 @@ class=trainers.cross_entropy_trainer.CrossEntropyTrainer
 decoders=[<decoder>]
 l2_weight=1.0e-8
 clip_norm=1.0
-optimizer_getter=<adadelta>
+optimizer=<adadelta>
 
 [adadelta]
-class=trainers.optimizers.adadelta_optimizer
+class=tf.train.AdadeltaOptimizer
 learning_rate=1.0e-4
 epsilon=1.0e-6
 rho=0.95
