+++ tests/transformer.ini
@@ -76,19 +76,19 @@ n_heads_enc=2
 class=trainers.cross_entropy_trainer.CrossEntropyTrainer
 decoders=[<decoder>]
 optimizer_getter=<lazyadam_g>
-decay_function=<decay>
 
-[decay]
-class=trainers.lr_decay.transformer_decay
+[decayed_lr]
+class=functions.noam_decay
+learning_rate=0.2
 model_dimension=6
 warmup_steps=111
 
 [lazyadam_g]
-class=trainers.optimizers.lazy_adam_optimizer
+class=tf.contrib.opt.LazyAdamOptimizer
 beta1=0.9
 beta2=0.98
 epsilon=1.0e-9
-learning_rate=0.2
+learning_rate=<decayed_lr>
 
 [runner]
 class=runners.GreedyRunner
